refactor(viewer): restore the taint barrier and the code-health scores (#1889)
SonarCloud dropped the PR's Security Rating to B, and CodeScene flagged a
hotspot decline in viewer.js. Both are consequences of how the previous
commits were shaped, not of what they do.

**Security (jssecurity:S8476, the B rating).** Extracting the three cell
fetches into fetchScenarioPayloads left the RECORDING_SLUG_RE barrier behind
in loadScenario, one function away from the request sinks it guards — so the
URL segments read as unvalidated at the point of use. The check now lives
beside the fetches it protects and the helper returns null for a hash that
is not a legal cell. A barrier one function away is a barrier both a reader
and a taint analyser have to go looking for.

**Complexity.** All advisory, all mechanical:

  - loadScenario: catalogEntryFor + a one-line recipe lookup take its
    cognitive complexity back under the limit (javascript:S3776)
  - renderRecordingHistory: 8 positional parameters → one options object
    (javascript:S107); its intro line and error banner extracted, which
    recovers the viewer.js hotspot
  - handleScenarioDetail: URL + profile validation moved into
    parseScenarioTarget, so every rejection is in one place and the handler
    reads as the happy path
  - catalog.go: the three annotation passes (measurement, pipeline,
    display-state) moved verbatim into catalog_annotate.go — catalog.go is
    now the skeleton only, 179 lines instead of 490
  - "manifest.json" → manifestFileName (go:S1192)
  - evidenceLink's three string arguments → evidenceReference
  - appendEvidenceRows' compound guard split
  - the two over-complex tests split into named assertion helpers

No behaviour change: the mutation proofs were re-run after the refactor and
still bite — dropping ?profile= from the cell fetches fails
viewer.profile.test.js, and neutralising the isolation guard fails
TestDesktopResultRefusesACLIRecording ("a Desktop result linked the CLI
recording \"a-cli\"").
