+++ tools/onboarding-factory/internal/viewer/scenarios.go
@@ -29,13 +29,24 @@ func (s *Server) handleScenariosList(w http.ResponseWriter, r *http.Request) {
 	writeJSON(w, s.store().listScenarios())
 }
 
-func (s *Server) handleScenarioDetail(w http.ResponseWriter, r *http.Request) {
+// scenarioTarget is one validated cell request: the three URL segments, the
+// remaining path parts, and the execution profile the whole request is scoped
+// to.
+type scenarioTarget struct {
+	agent, subtree, id string
+	parts              []string
+	profile            matrix.ExecutionProfile
+}
+
+// parseScenarioTarget validates the URL segments and ?profile=, writing the
+// error response itself and reporting ok=false. Keeping every rejection in one
+// place is what lets the handler below read as the happy path.
+func parseScenarioTarget(w http.ResponseWriter, r *http.Request) (scenarioTarget, bool) {
 	// URL form: /api/scenarios/{agent}/{subtree}/{id}[/recordings[/{name}]]
-	rest := strings.TrimPrefix(r.URL.Path, "/api/scenarios/")
-	parts := strings.Split(rest, "/")
+	parts := strings.Split(strings.TrimPrefix(r.URL.Path, "/api/scenarios/"), "/")
 	if len(parts) < 3 {
 		http.Error(w, "usage: /api/scenarios/{agent}/{subtree}/{id}", http.StatusBadRequest)
-		return
+		return scenarioTarget{}, false
 	}
 	// filepath.Base reduces agent/id to a single path segment before the
 	// ^[a-z0-9][a-z0-9_-]*$ slug check below — a no-op for any value that
@@ -44,14 +55,16 @@ func (s *Server) handleScenarioDetail(w http.ResponseWriter, r *http.Request) {
 	// recognizes for the file reads several hops downstream (recDir,
 	// scenarioDir, ...), where a regex match alone doesn't visibly clear
 	// the taint (see shard.sanitizePathComponent for the same idiom).
-	agent, subtree, id := filepath.Base(parts[0]), parts[1], filepath.Base(parts[2])
-	if subtree != "scenarios" && subtree != "regressions" {
-		http.Error(w, "subtree must be 'scenarios' or 'regressions'", http.StatusBadRequest)
-		return
+	target := scenarioTarget{
+		agent: filepath.Base(parts[0]), subtree: parts[1], id: filepath.Base(parts[2]), parts: parts,
 	}
-	if !slugRE.MatchString(agent) || !slugRE.MatchString(id) {
+	if target.subtree != "scenarios" && target.subtree != "regressions" {
+		http.Error(w, "subtree must be 'scenarios' or 'regressions'", http.StatusBadRequest)
+		return scenarioTarget{}, false
+	}
+	if !slugRE.MatchString(target.agent) || !slugRE.MatchString(target.id) {
 		http.Error(w, "agent and id must match ^[a-z0-9][a-z0-9_-]*$", http.StatusBadRequest)
-		return
+		return scenarioTarget{}, false
 	}
 	// Execution profile (#1889). Absent ?profile= is the cli-local default, so
 	// every pre-existing viewer URL keeps its meaning; an unknown value is a
@@ -59,8 +72,18 @@ func (s *Server) handleScenarioDetail(w http.ResponseWriter, r *http.Request) {
 	profile, err := profileFromRequest(r)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
+		return scenarioTarget{}, false
+	}
+	target.profile = profile
+	return target, true
+}
+
+func (s *Server) handleScenarioDetail(w http.ResponseWriter, r *http.Request) {
+	target, ok := parseScenarioTarget(w, r)
+	if !ok {
 		return
 	}
+	agent, subtree, id, profile, parts := target.agent, target.subtree, target.id, target.profile, target.parts
 	store := s.store()
 	scenarioDir := store.scenarioDir(agent, subtree, id)
 	if !store.exists(scenarioDir) {
@@ -292,7 +315,7 @@ func buildLatestManifest(recDir string, d *ScenarioDetail, store RecordingStore)
 		return nil
 	}
 	m := &RecordingArchive{Name: filepath.Base(recDir), DaemonVersion: "dev"}
-	if b, err := os.ReadFile(filepath.Join(recDir, "manifest.json")); err == nil {
+	if b, err := os.ReadFile(filepath.Join(recDir, manifestFileName)); err == nil {
 		if err := json.Unmarshal(b, m); err != nil {
 			logViewerError("buildLatestManifest: malformed manifest.json in %s: %v", recDir, err)
 		}
